requirements.txt
DateTime==4.3
pathlib==1.0.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
zope.interface==5.4.0
